Delete danmaku file error.

+++ iina+/Utils/Danmaku.swift
@@ -89,7 +89,9 @@ class Danmaku: NSObject {
     func prepareBlockList() throws {
         guard let resourcePath = Bundle.main.resourcePath else { return }
         let targetPath = resourcePath + "/Danmaku/iina-plus-blockList.xml"
-        try FileManager.default.removeItem(atPath: targetPath)
+        if FileManager.default.fileExists(atPath: targetPath) {
+            try FileManager.default.removeItem(atPath: targetPath)
+        }
         switch Preferences.shared.dmBlockList.type {
         case .none:
             return
